combine strings and literals in one test file
